requirements.txt
# ─── Core Web Framework ───────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ─── Streamlit ────────────────────────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.31.0

# ─── LLM SDKs (both free tiers) ──────────────────────────────────────────────
groq>=0.9.0
google-generativeai>=0.5.0
google-genai>=1.0.0            # newer Gemini SDK (Vision / multimodal)

# ─── LangGraph (multi-agent orchestration) ────────────────────────────────────
langgraph>=0.1.0
langchain-core>=0.2.0

# ─── RAG / Embeddings / Vector Store ──────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0               # use faiss-gpu on CUDA machines
numpy>=1.26.0

# ─── Math (SymPy) ─────────────────────────────────────────────────────────────
sympy>=1.12

# ─── OCR (PaddleOCR primary, EasyOCR optional fallback) ────────────────────────────
opencv-python-headless>=4.9.0  # image preprocessing (used by OCR backends)
paddleocr>=3.4.0               # PaddleOCR 3.x — bundles its own inference runtime
                               # (no separate paddlepaddle install needed on Python≥3.12)
# easyocr>=1.7.1               # optional fallback (set OCR_BACKEND=easyocr to enable)

# ─── Speech-to-Text (faster-whisper — local, free, 2–4× faster than openai-whisper) ─────
faster-whisper>=1.0.0          # CTranslate2-based Whisper — int8 CPU / float16 GPU
imageio-ffmpeg>=0.5.0          # bundled ffmpeg binary — avoids PATH issues on Windows
soundfile>=0.12.0              # read WAV/FLAC into numpy
pydub>=0.25.1                  # audio manipulation utilities

# ─── Data / Serialisation ─────────────────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.0

# ─── Utilities ────────────────────────────────────────────────────────────────
httpx>=0.27.0
aiofiles>=23.2.0
tiktoken>=0.7.0                # used by Whisper internals

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0                  # for FastAPI TestClient
